requirements.txt
Flask==3.0.3
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
tensorflow-cpu==2.13.0
keras==2.13.1
